refactor : buySatisfy Int형에서 Float형으로 변환

+++ service-product/module-batch/src/main/java/com/github/msafriends/modulebatch/processor/ElevenStreetProductItemProcessor.java
@@ -40,7 +40,7 @@ public class ElevenStreetProductItemProcessor implements ItemProcessor<ElevenStr
             item.getSalePrice(),
             REPLACED_QUANTITY,
             item.getDelivery(),
-            item.getBuySatisfy(),
+            (float)item.getBuySatisfy(),
             extractIntValue(benefit, "Discount"),
             extractIntValue(benefit, "Mileage"),
             extractAgeLimit(item.getMinorYn()),
